requirements.txt
aiogram==2.16.0
python-dotenv==0.19.1
telethon==1.23.0